reworked tkui, defined commands
Reworked the tkui to be used without
the UI-Shield buttons as remote controll.
The UI-Shield will be implemented as remote
interface getting the lcd text send to and
sending the button presses to the controller.

Reworked Plotting but still bad performance,
Data Storage type will be chamged from
pandas.Dataframe to numpy.structured array.

New commands need to be implented on controller.